In RESTException.get_errors(), check the type of self.errors.
RESTException is instantiated in some places (e.g. in invenio-files-rest
and in invenio-s3) with a simple string as construction parameter,
rather than with a list of FieldError objects. Therefore the call to
FieldError.to_dict() fails, and the code crashes. This change makes it so
that, in such a case, the bare string is simply passed along.